+++ crates/symbolicator/src/config.rs
@@ -1,3 +1,4 @@
+use std::collections::BTreeMap;
 use std::env;
 use std::fs;
 use std::path::{Path, PathBuf};
@@ -59,6 +60,10 @@ pub struct Metrics {
     pub hostname_tag: Option<String>,
     /// A tag name to report the environment to, for each metric. Defaults to not sending such a tag.
     pub environment_tag: Option<String>,
+    /// A map containing custom tags and their values.
+    ///
+    /// These tags will be appended to every metric.
+    pub custom_tags: BTreeMap<String, String>,
 }
 
 impl Default for Metrics {
@@ -71,6 +76,7 @@ impl Default for Metrics {
             prefix: "symbolicator".into(),
             hostname_tag: None,
             environment_tag: None,
+            custom_tags: BTreeMap::new(),
         }
     }
 }